Use CLOCK_MONOTONIC
CLOCK_REALTIME reflects the current date and time, which can be adjusted
by the system (e.g. NTP synchronization or manual change by the
administrator).

We get the current time to measue elapsed time, so that should use
CLOCK_MONOTONIC instead of CLOCK_REALTIME.

Also get_time_seconds() is removed, as it can be replaced with
get_time_micros() for precision.

Signed-off-by: Joonyoung Shim <joonyoung.shim@42dot.ai>